Adds new set of assay data for all samples in a FacileDataSet

Once a FacileDataSet has been created and initialized, either via a low-level call to initializeFacileDataSet(), or a call to as.FacileDataSet() over a list of BiocAssayContainers, you can add more assays (i.e. RNA-seq, microarray, etc.) to the FacileDataSet using this function.

addFacileAssaySet(x, datasets, facile_assay_name,
  facile_assay_type = .assay.types,
  facile_feature_type = .feature.types,
  facile_assay_description = NULL, facile_feature_info,
  storage_mode = .storage.modes, chunk_rows = 5000,
  chunk_cols = "ncol", chunk_compression = 4, assay_name = NULL,
  warn_existing = FALSE)

Arguments

x
The FacileDataSet

datasets
list of ExpressionSet, SummarizedExperiment, or DGELists that have the new assay data across all of the datasets in x.

facile_assay_name
the name of the assay in the source dataset object

facile_assay_type
string indicating the assay_type ('rnaseq', 'affymetrix', etc.)

facile_feature_type
a string indicating the universe the features in this assay refer to, i.e. "entrez", "ensgid", "enstid", etc.

facile_feature_info
a data.frame with the required feature_info columns that describe the features in this assay. Please refer to the "Features" section of the FacileDataSet vignette for more complete description.

storage_mode
either "integer" or "numeric", maps to the storage.mode parameter in rhdf5::h5createDataset()

chunk_rows
the first entry in the chunk parameter in rhdf5::h5createDataset() (integer)

chunk_cols
the second entry in the chunk parameter in rhdf5::h5createDataset(). If this is "ncol", it is set to the number of columns in each of the internal dataset matrices being added.

chunk_compression
the level parameter in rhdf5::h5createDataset()

assay_name
the assay name in the data containers provided in the datasets list.

facie_assay_description
a string that allows the caller to provide a "freeform" description of the assay (platform, protocol, whatever).

Value

a tibble subset of facile_feature_info that indicates the new features that were added to the internal feature_info_tbl.

Details

Note that you cannot add assay data piecemeal. That is to say, you can not call this function once to add copynumber data (addFacileAssaySet(..., facile_assay_type = "cnv") to a subset of samples and later call this function again to add copynumber to the rest of the samples. The function will throw an error if facile_assay_type