Encodes column(s) from pData into character values

This function is not exported, and should only be called from within the as.EAVtable() function because we rely on validity checks that are happening there.

eav_encode(dat, covariate_def, varname)

Arguments

dat
the vector to values to encode into an EAV table

covariate_def
the single-list-definition of this covariate

vname
the name of the attribute column in the eav table

Value

a four-column data.frame (dataset,sample_id,variable,value) with the encoded covariate into a single value column.