Fetch data from single assay of choice

Fetch data from single assay of choice

# S3 method for FacileDataSet
fetch_assay_data(x, features, samples = NULL,
  assay_name = default_assay(x), normalized = FALSE,
  as.matrix = FALSE, ..., subset.threshold = 700, aggregate = FALSE,
  aggregate.by = "ewm", verbose = FALSE)

# S3 method for facile_frame
fetch_assay_data(x, features, samples = NULL,
  assay_name = NULL, normalized = FALSE, as.matrix = FALSE, ...,
  subset.threshold = 700, aggregate = FALSE, aggregate.by = "ewm",
  verbose = FALSE)

normalize.assay.matrix(vals, feature.info, sample.info, fds, log = TRUE,
  prior.count = 1, batch = NULL, main = NULL,
  maintain.rowmeans = TRUE, verbose = FALSE, ...)

Arguments

x
A FacileDataSrote object.

features
a feature descriptor (data.frame with assay and feature_id columms)

samples
a sample descriptor to specify which samples to return data from.

assay_name
the name of the assay to fetch data from. Defaults to the value of default_assay() for x. Must be a subset of assay_names(x).

normalized
return normalize or raw data values, defaults to raw

as.matrix
by default, the data is returned in a long-form tbl-like result. If set to TRUE, the data is returned as a matrix.

...
parameters to pass to normalization methods

subset.threshold
sometimes fetching all the genes is faster than trying to subset. We have to figure out why that is, but I've previously tested random features of different lengths, and around 700 features was the elbow.

aggregate.by
do you want individual level results or geneset scores? Use 'ewm' for eigenWeightedMean, and that's all.

Value

A tibble (lazy or not) with assay data.

Removing Batch Effects

We leverage limma's removeBatchEffect functionality, with a simplified interface. The batch parameter replaces batch, batch2, and covariates. The design parameter is replaced with main.

All these parameters must be characters, which reference columns that have been passed down into the samples frame, or ones that we can pull out of the facile data store and spank onto the samples frame.

We'll use these parameters to build a model.matrix with main and batch effect and follow the use of removeBatchEffect as outlined in the post linked to below to pull the design matrix apart and call the function with the corresponding design and covariates parameters:

https://support.bioconductor.org/p/83286/#83287

Setting the batch.scale parameter to TRUE (the default), ensures that the rowMeans of the returned data matrix are the same as the original dataset.

See also

Examples

samples <- exampleFacileDataSet() %>%
  filter_samples(indication == "BLCA", sample_type == "tumor")
features <- c(PRF1='5551', GZMA='3001', CD274='29126')
dat <- with_assay_data(samples, features, normalized = TRUE, batch = "sex")
dat <- with_assay_data(samples, features, normalized = TRUE,
                       batch = c("sex", "RIN"), normalized = TRUE)

#> Error in with_assay_data(samples, features, normalized = TRUE, batch = c("sex",     "RIN"), normalized = TRUE): formal argument "normalized" matched by multiple actual arguments

dat <- with_assay_data(samples, features, normalized = TRUE,
                       batch = c("sex", "RIN"), main = "treatment")

#> Error: Can't find columns `treatment`, `RIN` in `.data`.