Path to the meta information YAML file

Lots of useful information is stored in a FacileDataSet's meta.yaml file. This function returns all of that in a list-of-lists

A FacileDataSet can contain assay data from different "datasets" (such as different cancer indications from the TCGA). This functions returns description and URL information that describes these datasets in more detail, which is specified in the FacileDataSets meta.yaml file.

meta_file(x)

meta_info(x, fn = meta_file(x))

# S3 method for FacileDataSet
default_assay(x)

dataset_definitions(x, as.list = TRUE)

Arguments

x
A FacileDataSet

fn
The path to the meta.yaml file.

as.list
boolean, if FALSE (default) returns a list, otherwise summarizes results into a tibble.

Value

The meta.yaml file parsed into a list-of-lists representation

meta information about the datasets in x as a list or tibble

See also