Creates a filter expression to select samples based on value of a covariate

This leverages dplyr's standard (vs non-standard) evaluation mojo. There is likely a cleaner way to do this, but to be honest I still find the interp stuff rather confusing

parse_sample_criterion(variable, value)

Arguments

variable
the name of the variable to look for in the sample_covariate variable column

value
character vector of values for the variable that you want your samples to have.

Value

a

See also