Saves custom sample covariates to a FacileDataSet

Saves custom sample covariates to a FacileDataSet

save_custom_sample_covariates(
  x,
  annotation,
  name = NULL,
  class = "categorical",
  custom_key = Sys.getenv("USER"),
  file.prefix = "facile",
  sample_filter_critera = NULL
)

Arguments

x
the FacileDataSet

annotation
the annotation table of covariate values to a sample-descriptor-like table

name
the variable name of the covariate

custom_key
the custom key (likely userid) for the annotation

file.prefix
Vincent uses this

sample_filter_criteria
optional list of filtering criteria that were used to drill down into the samples we have the annotatino data.frame for TODO: Figure out how to encode sample_filter_criteria into serialized (JSON) annotation file